Add DescSet
Introduce the type `DescSet`, a set of descriptors with convenient accessors to look up descriptors by name.

This type is the successor to `DescriptorSet` and registries, and combines the features of both.
